Move calculated_size utils to size utils
